+++ derive_helpers/src/utilities/mod.rs
@@ -5,12 +5,10 @@
  * Licensed under the MIT license
  */
 pub mod ident;
-pub mod renaming;
 pub mod pick;
 pub mod pickable_table_derive_attr;
 
 pub use ident::*;
-pub use renaming::*;
 pub use pick::*;
 pub use pickable_table_derive_attr::*;
